requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.8
packaging==24.1
pycountry==24.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
zope.event==5.0
zope.interface==7.0.3
